Refs #12288 - handle the missing proxy action when the proxy action fails in planning phase
